fixes at get neighbors

+++ codebase/rust/src/globe/coordinates_system/surface_point.rs
@@ -2,8 +2,8 @@
 use godot::{classes::{Area3D, IArea3D, MeshInstance3D, PhysicsRayQueryParameters3D, World3D}, prelude::*};
 use crate::{globe::territories::territory::TerritoryId, troops::troop::Troop};
 
-type Latitude = i32;
-type Longitude = i32;
+type Latitude = i16;
+type Longitude = i16;
 pub type Coordinates = (Latitude, Longitude);
 
 #[derive(Debug, Clone)]
